Added Sprites!
Moved the entity example in main to it's own struct. The Render function (called from main with the application instance) now only needs the sprite as it's argument (from needing the texture (or color) and the anchors transform)!

I also cleaned up the Main.cpp file to not be so cluttered.