Added support to disable the Viewport
Previously the main window was rendered to a texture rendered in a viewport window displayed by ImGui.
It's now possible to disable that and just render a "normal" GLFW window with a function call before creating the application window.
(LFW::Viewport::Init(true) to use the Viewport or LFW::Viewport::Init(false) to use the original GLFW window).